Add review week to schedule
As per our discussion yesterday, this coming week will be a review week.
This will give us a chance to play with the code exercises, re-read
challenging sections, and catch up on skipped work, and generally let
the material sink in.

I put myself as the facilitator, but if someone else want to take the
reins, feel free to change that.

If there's no dissent, I'll merge this PR later tonight.