Add basic README for LYSE
This is basically a copy of the README from *Elixir in Action* with some
small changes for the new book. I've removed the meeting details, leader
column, and member list (we'll decide on that later on), and I've updated
the schedule for my best guess of our reading pace.

Since this is so closely tied to our next book, *Designing for
Scalability with Erlang/OTP*, I included a link at the bottom of the
schedule.

When we have a solid member list and decide when/where/how we're
meeting, we can finish this up.