Add root README
Now that we have more than one book here, the root README is more
useful. It explains the purpose of the group and provides a list of the
books we've read.

I assume that someone would be reading it because 1) they want to find
out more about what we're doing/how we're organizing the reading group,
or 2) they're in the group and want quick access to the relevant book. I
hope that this provides for both cases.

When we settle on the dates for LYSE and DFS, I'll update the chart.