Update reading schedule for split ch12
We decided last week to split Chapter 12 into two sections since it's
such a long chapter. For the April 10th meeting, read up to 12.2; the
next meeting, we'll go from that section to the end of the chapter.

Since we only had three attendees last meeting, we'll also talk about
Chapter 11 on April 10th if there's still material to discuss.

For the meeting leader, I just kept rotating names alphabetically. Let
me know if that's a problem!